+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+11.10.2013
+------------------
+* Added params to users action for store custom information
+* Migrated all modules to new Crud system
+* Updated `users_action` table (added `params` field)
+
 08.10.2013
 ------------------
 * Added CHANGELOG.md file (this file)
@@ -21,7 +27,6 @@ CHANGELOG
 05.09.2013
 ------------------
 * Replaced Gliphicons with Awesome
-
 
 05.08.2013
 ------------------